fix movement in game over screen
