+++ Juego/Assets/Scripts/Enemy/GameOver.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 166c71d6f34d79f488648c6085712ced
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
